requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
sentence-transformers>=2.7.0
pyyaml==6.0.1
requests==2.31.0
pydantic>=2.8.0
tqdm==4.66.1
typing-extensions>=4.10.0
beautifulsoup4==4.12.2
html2text==2020.1.16
